Move DVC domain services API to the domain sub-tree
