+++ src/nautilus_librarian/typer/commands/workflows/actions/generate_base_images_action.py
@@ -1,5 +1,5 @@
 from nautilus_librarian.domain.file_locator import file_locator
-from nautilus_librarian.mods.dvc.domain.dvc_services_api import DvcServicesApi
+from nautilus_librarian.domain.dvc_services_api import DvcServicesApi
 from nautilus_librarian.mods.dvc.domain.utils import (
     extract_added_and_modified_files_from_dvc_diff,
 )
